ifobject: new serialization API

+++ ifobject/include/ifobject/types.hpp
@@ -69,7 +69,7 @@ struct IFOpInfo;
 // other types
 typedef std::string IFObjectID;
 typedef std::string IFParamID;
-typedef int IFIDNum;
+typedef Int64 IFIDNum;
 typedef int IFEventType;
 typedef std::set<Ionflux::ObjectBase::IFObject*> IFObjectSet;
 typedef std::vector<Ionflux::ObjectBase::IFObject*> IFObjectVector;
